+++ examples/rustapi_module/tests/test_subclassing.py
@@ -7,10 +7,11 @@ PYPY = platform.python_implementation() == "PyPy"
 
 class SomeSubClass(Subclassable):
     def __str__(self):
-        return "Subclass"
+        return "SomeSubclass"
 
 
 def test_subclassing():
     if not PYPY:
         a = SomeSubClass()
-        assert str(a) == "Subclass"
+        assert str(a) == "SomeSubclass"
+        assert type(a) is SomeSubClass