+++ pyo3-derive-backend/src/pymethod.rs
@@ -193,7 +193,7 @@ pub fn impl_wrap_new(cls: &syn::Type, spec: &FnSpec<'_>) -> TokenStream {
     quote! {
         #[allow(unused_mut)]
         unsafe extern "C" fn __wrap(
-            subcls: *mut pyo3::ffi::PyTypeObject,
+            subtype: *mut pyo3::ffi::PyTypeObject,
             _args: *mut pyo3::ffi::PyObject,
             _kwargs: *mut pyo3::ffi::PyObject) -> *mut pyo3::ffi::PyObject
         {
@@ -206,7 +206,7 @@ pub fn impl_wrap_new(cls: &syn::Type, spec: &FnSpec<'_>) -> TokenStream {
 
                 let _result = pyo3::derive_utils::IntoPyNewResult::into_pynew_result(#body)?;
                 let initializer = pyo3::PyClassInitializer::from(_result);
-                let cell = initializer.create_cell_from_subtype(_py, subcls)?;
+                let cell = initializer.create_cell_from_subtype(_py, subtype)?;
                 Ok(cell as *mut pyo3::ffi::PyObject)
             })
         }